requirements.txt
# RAG Pipeline - zależności
# Vector DB
qdrant-client[fastembed]>=1.13.0

# LLM i embeddingi
anthropic>=0.40.0
httpx>=0.27.0

# Chunking
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# Ewaluacja
ragas>=0.2.0

# Frontend
streamlit>=1.35.0
plotly>=5.20.0

# Narzędzia
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=2.2.0
tqdm>=4.66.0
